Fix cast from HANDLE to descriptor in the WinAPI driver.
Avoids truncation to 32-bit for 64-bit builds. Also improves robustness of the directory watcher shutdown procedure.

+++ source/eventcore/drivers/winapi/events.d
@@ -142,6 +142,6 @@ final class WinAPIEventDriverEvents : EventDriverEvents {
 
 	private static HANDLE idToHandle(EventID event)
 	@trusted {
-		return cast(HANDLE)cast(int)event;
+		return cast(HANDLE)cast(size_t)event;
 	}
 }
